+++ metadata.rb
@@ -3,7 +3,7 @@ maintainer        "Opscode, Inc."
 maintainer_email  "matt@opscode.com"
 license           "Apache 2.0"
 description       "Provides the shared messaging configuration for Chef for OpenStack."
-version           "7.0.0"
+version           "7.1.0"
 
 recipe "server", "Installs and configures server packages for messaging queue used by the deployment."
 recipe "rabbitmq-server", "Installs and configures RabbitMQ and is called via the server recipe"